requirements.txt
pydantic-ai
python-telegram-bot
python-dotenv
httpx
pytest
pytest-asyncio
ruff
